requirements.txt
fastAPI
uvicorn
python-multipart
Pillow
python-dotenv
sqlalchemy